Move slack and utils tests from test_views to test_services
